@@ -2,7 +2,7 @@ import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { MatCardModule } from '@angular/material/card';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
@@ -17,30 +17,24 @@ import { CodeBlockComponent } from './code-block/code-block.component';
 import { OutputDisplayComponent } from './output-display/output-display.component';
 import { SOCKET, socket } from './socket';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    BlockLibraryComponent,
-    CanvasComponent,
-    CodeBlockComponent,
-    OutputDisplayComponent,
-  ],
-  imports: [
-    BrowserModule,
-    BrowserAnimationsModule,
-    FormsModule,
-    HttpClientModule,
-    MatCardModule,
-    MatFormFieldModule,
-    MatProgressSpinnerModule,
-    MatSelectModule,
-    MatSnackBarModule,
-    MatTooltipModule
-  ],
-  providers: [
-    {provide: SOCKET, useValue: socket},
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        BlockLibraryComponent,
+        CanvasComponent,
+        CodeBlockComponent,
+        OutputDisplayComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BrowserAnimationsModule,
+        FormsModule,
+        MatCardModule,
+        MatFormFieldModule,
+        MatProgressSpinnerModule,
+        MatSelectModule,
+        MatSnackBarModule,
+        MatTooltipModule], providers: [
+        { provide: SOCKET, useValue: socket },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 
 export class AppModule { }
