@@ -1,4 +1,4 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { MatCardModule } from '@angular/material/card';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { TestBed } from '@angular/core/testing';
@@ -9,24 +9,24 @@ import { CanvasComponent } from './canvas/canvas.component';
 import { MockSocket } from './mock-socket';
 import { OutputDisplayComponent } from './output-display/output-display.component';
 import { SOCKET } from './socket';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('AppComponent', () => {
   beforeEach(() => TestBed.configureTestingModule({
     declarations: [
-      AppComponent,
-      BlockLibraryComponent,
-      CanvasComponent,
-      OutputDisplayComponent,
+        AppComponent,
+        BlockLibraryComponent,
+        CanvasComponent,
+        OutputDisplayComponent,
     ],
+    imports: [MatCardModule,
+        MatSnackBarModule],
     providers: [
-      { provide: SOCKET, useClass: MockSocket },
-    ],
-    imports: [
-      HttpClientTestingModule,
-      MatCardModule,
-      MatSnackBarModule,
-    ],
-  }));
+        { provide: SOCKET, useClass: MockSocket },
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting(),
+    ]
+}));
 
   it('should create the app', () => {
     const fixture = TestBed.createComponent(AppComponent);
